+++ core/state.go
@@ -2,7 +2,7 @@ package core
 
 // State defines the state of the editor.
 type State struct {
-	Line   int
+	Line   int64
 	Cursor int
 	Bytes  []byte
 	Size   int
